refactor(unifi-protect): clarify media request lifecycle
